requirements.txt
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
opencv-python==4.0.0.21
numpy==1.14.5
tensorflow-gpu==1.10.0
